+++ test/default/test_defaults.py
@@ -4,13 +4,12 @@ from sklearn.datasets import load_iris, fetch_california_housing, load_breast_ca
 from sklearn.model_selection import train_test_split
 import pandas as pd
 from flaml import AutoML
-from flaml.automl.default import (
-    portfolio,
+from flaml.default import (
     preprocess_and_suggest_hyperparams,
     suggest_hyperparams,
     suggest_learner,
 )
-from flaml.automl.default import regret
+from flaml.default import portfolio, regret
 
 
 def test_greedy_feedback(path="test/default", strategy="greedy-feedback"):
@@ -147,7 +146,7 @@ def test_suggest_regression():
 
 
 def test_rf():
-    from flaml.automl.default import RandomForestRegressor, RandomForestClassifier
+    from flaml.default import RandomForestRegressor, RandomForestClassifier
 
     X_train, y_train = load_breast_cancer(return_X_y=True, as_frame=True)
     rf = RandomForestClassifier()
@@ -165,7 +164,7 @@ def test_rf():
 
 
 def test_extratrees():
-    from flaml.automl.default import ExtraTreesRegressor, ExtraTreesClassifier
+    from flaml.default import ExtraTreesRegressor, ExtraTreesClassifier
 
     X_train, y_train = load_iris(return_X_y=True, as_frame=True)
     classifier = ExtraTreesClassifier()
@@ -183,7 +182,7 @@ def test_extratrees():
 
 
 def test_lgbm():
-    from flaml.automl.default import LGBMRegressor, LGBMClassifier
+    from flaml.default import LGBMRegressor, LGBMClassifier
 
     X_train, y_train = load_breast_cancer(return_X_y=True, as_frame=True)
     classifier = LGBMClassifier(n_jobs=1)
@@ -203,7 +202,7 @@ def test_lgbm():
 
 
 def test_xgboost():
-    from flaml.automl.default import XGBRegressor, XGBClassifier
+    from flaml.default import XGBRegressor, XGBClassifier
 
     X_train, y_train = load_breast_cancer(return_X_y=True, as_frame=True)
     classifier = XGBClassifier(max_depth=0)
